backend/requirements.txt
asgiref==3.7.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
Django==4.2.5
django-cors-headers==4.2.0
djangorestframework==3.14.0
drf-yasg==1.21.7
google-api-core==2.12.0
google-auth==2.23.2
google-cloud-vision==3.4.4
googleapis-common-protos==1.60.0
grpcio==1.59.0
grpcio-status==1.59.0
idna==3.4
inflection==0.5.1
packaging==23.1
proto-plus==1.22.3
protobuf==4.24.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
sqlparse==0.4.4
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.5
